1.8.0 - 151 //  Feedback + Soft Update

+++ data/src/androidTest/java/ca/bc/gov/data/local/BaseDataBaseTest.kt
@@ -107,7 +107,7 @@ abstract class BaseDataBaseTest {
         text: String = "Sample Comment",
         entryTypeCode: String = "Med",
         parentEntryId: String = UUID.randomUUID().toString(),
-        version: Int = 0,
+        version: Long = 0L,
         createdDateTime: Instant = Instant.now(),
         createdBy: String = "Created By Pinakin",
         updateDateTime: Instant = Instant.now(),
